typography token renaming
